Refactor exceptions into chatty.exception package
