+++ src/main/java/chatty/exception/ChattyChatBotException.java
@@ -1,4 +1,4 @@
-package chatty;
+package chatty.exception;
 
 public class ChattyChatBotException extends Exception {
