+++ text-ui-test/runtest.sh
@@ -13,7 +13,7 @@ then
 fi
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src -Xlint:none -d ../bin ../src/main/java/chatty/*.java ../src/main/java/chatty/task/*.java ../src/main/java/chatty/util/*.java
+if ! javac -cp ../src -Xlint:none -d ../bin ../src/main/java/chatty/*.java ../src/main/java/chatty/task/*.java ../src/main/java/chatty/util/*.java ../src/main/java/chatty/exception/*.java
 then
     echo "********** BUILD FAILURE **********"
     exit 1
